Code refactoring

+++ Firmware/src/Config.h
@@ -97,9 +97,9 @@ struct ConfigObject
   uint8_t HumidityHysterisis;
 
   /**
-   * @brief Voltage set point below which the AC power supply will automatically connect
+   * @brief Battery capacity set point below which the AC power supply will automatically connect
    */
-  float AcBackupPoint;
+  uint8_t AcBackupPoint;
 };
 
 ConfigObject getDefaultConfig();
